Added button and link

+++ src/classes/ViewMeetingsManager.php
@@ -51,6 +51,7 @@ class ViewMeetingsManager {
             echo '              <th>Time</th>' . "\r\n";
             echo '              <th>Recurring</th>' . "\r\n";
             echo '              <th>Delete?</th>' . "\r\n";
+	    echo '              <th>Edit</th>' . "\r\n";
             echo '      </tr>' . "\r\n";
             echo '  </thead>' . "\r\n";
             echo '  <tbody>' . "\r\n";
@@ -81,9 +82,11 @@ class ViewMeetingsManager {
                     echo '<td>' . "<a class='home_page_link' onclick='return confirm(\"Are you sure?\")'href='cancel_reservation.php?reservation_id=" . $row['_id']  . "'>Delete</a>" . ' [' ."<a class='home_page_link' onclick='return confirm(\"This will remove all reservations in the recurrence. Are you sure this is what you want?\")'href='cancel_reservation.php?recurrence_id=" . $row['recurrence_id']  . "'>All</a>". ']</td>' . "\r\n";
                 } else {
                     echo '<td>' . "<a class='home_page_link' onclick='return confirm(\"Are you sure?\")'href='cancel_reservation.php?reservation_id=" . $row['_id']  . "'>Delete</a>" . '</td>' . "\r\n";
-
+	
                 }
+		echo '<td>' . "<a class='home_page_link' onclick='return confirm(\"Are you sure?\")'href='edit_reservation.php?reservation_id=" . $row['_id']  . "'>Edit</a>" . '</td>' . "\r\n";
      	        echo '      </tr>' . "\r\n";
+
             }
 
             echo '  </tbody>' . "\r\n";
